+++ src/main/java/fr/imie/kitmeal/controllers/UserController.java
@@ -41,9 +41,9 @@ public class UserController {
     public ModelAndView findAllUsers(HttpSession session,
             HttpServletRequest request) {
         List<UserBean> beans = userService.findAllUsers();
-        if(session.getAttribute("user") != null){
+        if (session.getAttribute("user") != null) {
             return new ModelAndView("/contacts/contacts.jsp", "bean", beans);
-        }else{
+        } else {
             return new ModelAndView("redirect:/app/log");
         }
     }
@@ -54,21 +54,26 @@ public class UserController {
         AddressBean address = addressService.createAddress(bean.getAddress());
         userService.createUser(bean);
 
-        return new ModelAndView("redirect:/app/users");
+        if (session.getAttribute("user") != null) {
+            return new ModelAndView("redirect:/app/users");
+        } else {
+            return new ModelAndView("redirect:/app/log");
+        }
+
     }
 
     @RequestMapping(value = "/create", method = RequestMethod.GET)
     public ModelAndView showCreateUser(HttpSession session,
             HttpServletRequest request) {
         UserBean bean = new UserBean();
-        UserBean user = (UserBean)session.getAttribute("user");
-        if(session.getAttribute("user") != null){
-            if((user.getRole() == "admin")){
-            return new ModelAndView("/contacts/form.jsp", "bean", bean);
-            }else{
-              return new ModelAndView("redirect:/app/users");  
+        UserBean user = (UserBean) session.getAttribute("user");
+        if (session.getAttribute("user") != null) {
+            if (("admin".equals(user.getRole()))) {
+                return new ModelAndView("/contacts/form.jsp", "bean", bean);
+            } else {
+                return new ModelAndView("redirect:/app/users");
             }
-        }else{
+        } else {
             return new ModelAndView("redirect:/app/log");
         }
     }
@@ -76,25 +81,29 @@ public class UserController {
     @RequestMapping(value = "/update/{idUser}/{bean}", method = RequestMethod.POST)
     public ModelAndView updateUser(HttpSession session, UserBean bean,
             @PathVariable Integer idUser, HttpServletRequest request) {
+        if (session.getAttribute("user") != null) {
+            addressService.updateAddress(bean.getAddress().getIdAddress(), bean.getAddress());
+            userService.updateUser(idUser, bean);
 
-        addressService.updateAddress(bean.getAddress().getIdAddress(), bean.getAddress());
-        userService.updateUser(idUser, bean);
+            return new ModelAndView("redirect:/app/users");
+        } else {
+            return new ModelAndView("redirect:/app/log");
+        }
 
-        return new ModelAndView("redirect:/app/users");
     }
 
     @RequestMapping(value = "/update/{idUser}", method = RequestMethod.GET)
     public ModelAndView showUpdateUser(HttpSession session,
             @PathVariable Integer idUser, HttpServletRequest request) {
         UserBean bean = userService.findUser(idUser);
-        UserBean user = (UserBean)session.getAttribute("user");
-        if(session.getAttribute("user") != null){
-            if((user.getRole() == "admin")){
+        UserBean user = (UserBean) session.getAttribute("user");
+        if (session.getAttribute("user") != null) {
+            if (("admin".equals(user.getRole())) || user.getIdUser() == bean.getIdUser()) {
                 return new ModelAndView("/contacts/update.jsp", "bean", bean);
-            }else{
-                return new ModelAndView("redirect:/app/users");  
+            } else {
+                return new ModelAndView("redirect:/app/users");
             }
-        }else{
+        } else {
             return new ModelAndView("redirect:/app/log");
         }
     }
@@ -110,9 +119,19 @@ public class UserController {
     @RequestMapping(value = "/remove/{idUser}", method = RequestMethod.GET)
     public ModelAndView removeUser(HttpSession session, @PathVariable Integer idUser,
             HttpServletRequest request) {
-        userService.removeUser(idUser);
+        UserBean user = (UserBean) session.getAttribute("user");
 
-        return new ModelAndView("redirect:/app/users");
+        if (session.getAttribute("user") != null) {
+            if (("admin".equals(user.getRole()))) {
+                userService.removeUser(idUser);
+                return new ModelAndView("redirect:/app/users");
+            } else {
+                return new ModelAndView("redirect:/app/users");
+            }
+        } else {
+            return new ModelAndView("redirect:/app/log");
+        }
+
     }
 
 }
